requirements.txt
# Core
google-generativeai>=0.5.0
anthropic>=0.25.0
openai>=1.30.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0

# RAG & Vector DB
chromadb>=0.5.0
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
cohere>=5.0.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.7.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
kaleido>=0.2.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
markitdown>=0.0.1

# Web & API
httpx>=0.27.0
aiohttp>=3.9.0
fastapi>=0.111.0
uvicorn>=0.29.0

# Observability
structlog>=24.1.0
prometheus-client>=0.20.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0

# Dependency Injection & Resilience
dependency-injector>=4.41.0
tenacity>=8.2.0

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0

# Development
black>=24.3.0
ruff>=0.4.0
mypy>=1.10.0
pre-commit>=3.7.0
